requirements.txt
fastapi~=0.128.0
SQLAlchemy~=2.0.46
alembic~=1.18.3
pydantic~=2.12.5
pydantic-settings~=2.12.0
PyJWT~=2.11.0
passlib~=1.7.4
starlette~=0.50.0